add .env, move service interface into service, add pino logger
