+++ .env
@@ -1,0 +1,4 @@
+PORT=3001
+
+TOKEN_SECRET=my_secret_key
+TOKEN_EXP=4h
