+++ package.json
@@ -18,6 +18,7 @@
     "@types/node": "^13.13.2",
     "bcrypt": "^4.0.1",
     "bcryptjs": "^2.4.3",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "express-jwt": "^5.3.3",
     "helmet": "^3.22.0",
@@ -25,6 +26,7 @@
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.15",
     "morgan": "^1.10.0",
+    "pino": "^6.2.1",
     "uuid": "^8.0.0"
   },
   "devDependencies": {
@@ -36,6 +38,7 @@
     "@types/helmet": "0.0.47",
     "@types/jsonwebtoken": "^8.3.9",
     "@types/lodash": "^4.14.150",
+    "@types/pino": "^6.0.1",
     "@types/uuid": "^7.0.3",
     "awesome-typescript-loader": "^5.2.1",
     "source-map-loader": "^0.2.4",
