--- src/config.ts
@@ -1,7 +1,0 @@
-export const pathToDBFile = 'src/database.json';
-export const serverPort = process.env.PORT || 3001;
-export const secretKey = 'my secret key';
-export const tokenConfig = {
-  expiresIn: '4h'
-};
-
