--- src/database/database.models.ts
@@ -1,6 +1,0 @@
-export interface DatabaseInterface {
-  findAll(selector: string): Promise<[]>;
-  add(selector: string, data: {}): Promise<any>;
-  find(selector: string, query: object): Promise<any>;
-  remove(selector: string, query: object): Promise<any>;
-}
