+++ src/operation/operation.models.ts
@@ -5,9 +5,3 @@ export interface Operation {
   sum: number;
   currency: Currency;
 }
-
-export interface OperationInterface {
-  getData(): Promise<Operation[]>;
-  getDataById(selector: string, id: string): Promise<Operation>;
-  addData(newOperation: Operation): Promise<Operation>;
-}
